refactor: moved memory metrics and middelware to own file
